fix things
fetchPacScript is now always called on install, and is now only for
network fetch. saving to the cache is now callers' responsibility as
well.
the options set/get api is modified to be more like chrome.storage, but
that isn't used because sync isn't really a good idea
configureDeclarativeContent is no longer automatically called, either,
and is the caller's responsibility